requirements.txt
accelerate==1.6.0
av==12.0.0
braceexpand==0.1.7
click==8.2.1
diffusers==0.27.0
dill==0.4.0
einops==0.8.1
filelock==3.18.0
fvcore==0.1.5.post20221221
glfw==2.9.0
gradio==4.44.1
huggingface-hub==0.22.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
imgui==2.0.0
ipdb==0.13.13
lmdb==1.7.3
lpips==0.1.4
matplotlib==3.10.5
numpy==1.24.1
omegaconf==2.3.0
opencv_python==4.10.0.84
pandas==2.3.1
phalp==0.1.3
Pillow==11.3.0
pip==24.2
psutil==7.0.0
PyOpenGL==3.1.0
pyrender==0.1.45
pyrootutils==1.0.4
pyspng==0.1.3
pytorch_lightning==2.5.2
pytorch_msssim==1.0.0
Requests==2.32.4
rich==14.1.0
scikit-image==0.25.2
scipy==1.15.3
setuptools==75.1.0
smplx==0.1.28
spconv_cu121==2.3.8
timm==1.0.19
tqdm==4.67.1
transformers==4.39.3
trimesh==4.6.8
visualize==0.5.1
wandb==0.21.1
webdataset==1.0.2
xformers==0.0.22.post7
yacs==0.1.8

pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@75ebeeaea0908c5527e7b1e305fbc7681382db47
chumpy @ git+https://github.com/mattloper/chumpy@9b045ff5d6588a24a0bab52c83f032e2ba433e17
detectron2 @ git+https://github.com/facebookresearch/detectron2@18f69583391e5040043ca4f4bebd2c60f0ebfde0